--- src/fides/api/models/manual_tasks/__init__.py
@@ -1,20 +1,0 @@
-from fides.api.models.manual_tasks.manual_task import ManualTask, ManualTaskReference
-from fides.api.models.manual_tasks.manual_task_config import (
-    ManualTaskConfig,
-    ManualTaskConfigField,
-)
-from fides.api.models.manual_tasks.manual_task_instance import (
-    ManualTaskInstance,
-    ManualTaskSubmission,
-)
-from fides.api.models.manual_tasks.manual_task_log import ManualTaskLog
-
-__all__ = [
-    "ManualTask",
-    "ManualTaskConfig",
-    "ManualTaskConfigField",
-    "ManualTaskInstance",
-    "ManualTaskLog",
-    "ManualTaskReference",
-    "ManualTaskSubmission",
-]
